fix(Geometry): fixed that `compute_rotation_matrix` returned none if there is 0 in `inliers`
There was a check that affine transformation is successful, that it's successful if inliers does not contain zeroes. However, we use this only to calculate inlines and xlines from coordinates, which is done preciesly enough.  Also, a small validation test was added to test that ilines and xlines are calculated from coordinates as intended